adding backend handling for lab hour submission
