+++ django_project/ta_system/static/ta_system/home.css
@@ -48,10 +48,10 @@
     height: 0!important;
 }
 
-#cs-lab-hour-form {
+#lab-hour-form {
     margin-top: 20px;
 }
 
-#cs-lab-hour-submit-button {
+#lab-hour-submit-button {
     margin: 30px 0 20px 0;
 }
