+++ django_project/ta_system/static/ta_system/lab-hour-settings.js
