+++ django_project/ta_system/templates/ta_system/bad_request.html
@@ -1,8 +1,10 @@
 {% extends "ta_system/base.html" %}
 {% load i18n static %}
 {% block content %}
-    <h1>400: Bad Request</h1>
-    <p>The wrong type of HTTP request was made.</p>
-    <p><strong>Expected:</strong> {{ expected_method }}</p>
-    <p><strong>Received:</strong> {{ received_method }}</p>
+    <div class="content-section">
+        <h1>400: Bad Request</h1>
+        <p>The wrong type of HTTP request was made.</p>
+        <p><strong>Expected:</strong> {{ expected_method }}</p>
+        <p><strong>Received:</strong> {{ received_method }}</p>
+    </div>
 {% endblock %}
