+++ django_project/ta_system/templates/ta_system/home.html
@@ -2,11 +2,19 @@
 {% load static %}
 {% block extra-head %}
     <link rel="stylesheet" type="text/css" href="{% static 'ta_system/home.css' %}" />
-    <script src="{% static 'ta_system/home.js' %}" type="module" ></script>
+    <script type="module" src="https://cdn.jsdelivr.net/npm/js-cookie@rc/dist/js.cookie.min.mjs"></script>
+    <script src="{% static 'ta_system/lab-hour-form.js' %}" type="module" ></script>
 {% endblock %}
 {% block content %}
     <div class="content-section">
-        <h2>Apply Here</h2>
-        <div id="cs-lab-hour-form"></div>
+        {% if user_has_submitted_application %}
+            <h2>Success! Your Application has been submitted.</h2>
+            <p>
+                To edit your account info or update your CS Lab availability, head on over to your <a href= "{% url 'profile' %}">profile</a> page.
+            </p>
+        {% else %}
+            <h2>Apply Here</h2>
+            <div id="lab-hour-form"></div>
+        {% endif %}
     </div>
 {% endblock content %}
